Added chart for cumulative scores (Augment code)
Prompt:
This is my squash score tracking application done with ClojureScript, React, Reagent and Re-frame.
I would like to work with you to extend some new features to it. I am running the build so
that development-build is running at http://localhost:8280/ and updated automatically.

First I would like to have a chart of score development during a year on the year page
which is coded in year.cljs. The chart should be after the score table and show how
cumulative points of each player (Roope, Kari and Niklas) has been developing during the year.
It should be line-chart with time (date) on X-axist and cumulative score on Y-axis.
Different players should be represented with lines of different color.